// apps/frontend/src/views/form/materials/types.ts
export enum MaterialType {
  AppInput = "AppInput",
  AppTextarea = "AppTextarea",
  AppNumber = "AppNumber",
  AppDatetime = "AppDatetime",
  AppSelect = "AppSelect",
  AppMultiSelect = "AppMultiSelect",
  AppUploadImage = "AppUploadImage",
  AppUploadFile = "AppUploadFile",
  AppCollapse = "AppCollapse",
}